+++ header.php
@@ -23,8 +23,30 @@
 </head>
 
 <body <?php body_class(); ?>>
-<div id="page" class="site">
 
-    <?php get_template_part( 'templates/site-header' ); ?>
+    <?php 
+	/**
+	 * Functions hooked in to ipress_before_header
+	 *
+	 * @hooked ipress_skip_links - 10
+	 */
+    do_action( 'ipress_before_header' ); ?>
 
-    <div id="content" class="site-content">
+    <header id="masthead" class="site-header" role="banner" <?php ipress_header_style(); ?>>
+        <div class="wrap">
+            <?php 
+            /**
+             * Functions hooked into honeycomb_header action
+             *
+             * @hooked ipress_site_branding            - 10
+             * @hooked ipress_primary_navigation       - 20
+            */
+            do_action( 'ipress_header' ); ?>
+        </div>
+    </header>
+
+    <?php do_action( 'ipress_before_content' ); ?>
+
+    <div id="content" class="site-content" tabindex="-1">
+
+		<?php do_action( 'ipress_content_top' );