+++ route/author.php
@@ -33,13 +33,7 @@
     
         <?php rewind_posts(); ?>
 
-        <?php while ( have_posts() ) : the_post(); ?>
-    
-            <?php get_template_part( 'templates/content' ); ?>
-
-        <?php endwhile; ?>
-
-        <?php the_posts_navigation(); ?>
+        <?php get_template_part( 'templates/loop' ); ?>
 
     <?php else: ?>
 
@@ -50,5 +44,5 @@
 	</main><!-- #main -->
 </div><!-- #primary -->
 
-<?php get_sidebar(); ?>
+<?php do_action( 'ipress_sidebar' ); ?>
 <?php get_footer(); ?>