+++ route/index.php
@@ -30,13 +30,7 @@
         <?php endif; ?>
         </header><!-- .page-header -->
 
-        <?php while ( have_posts() ) : the_post(); ?>
-    
-            <?php get_template_part( 'templates/content' ); ?>
-
-        <?php endwhile; ?>
-
-        <?php the_posts_navigation(); ?>
+        <?php get_template_part( 'templates/loop' ); ?>
 
     <?php else: ?>
     
@@ -47,5 +41,5 @@
 	</main><!-- #main -->
 </div><!-- #primary -->
 
-<?php get_sidebar(); ?>
+<?php do_action( 'ipress_sidebar' ); ?>
 <?php get_footer(); ?>