+++ templates/loop-excerpt.php
@@ -1,0 +1,16 @@
+<?php 
+
+/**
+ * iPress - WordPress Theme Framework                       
+ * ==========================================================
+ *
+ * Template for displaying the post loop excerpt
+ * 
+ * @package     iPress\Templates
+ * @link        http://ipress.uk
+ * @license     GPL-2.0+
+ */
+?>
+<section class="entry-summary">
+	<?php the_excerpt(); ?>
+</section><!-- .entry-summary -->