+++ templates/loop-sticky.php
@@ -1,0 +1,16 @@
+<?php 
+
+/**
+ * iPress - WordPress Theme Framework                       
+ * ==========================================================
+ *
+ * Template for displaying the post loop sticky header
+ * 
+ * @package     iPress\Templates
+ * @link        http://ipress.uk
+ * @license     GPL-2.0+
+ */
+?>
+<?php if ( is_sticky() && is_home() ) : ?>
+    <div class="sticky-post"></div>
+<?php endif;