+++ templates/skip-links.php
@@ -1,0 +1,15 @@
+<?php 
+
+/**
+ * iPress - WordPress Theme Framework                       
+ * ==========================================================
+ *
+ * Template for displaying skip links
+ * 
+ * @package     iPress\Templates
+ * @link        http://ipress.uk
+ * @license     GPL-2.0+
+ */
+?>
+<a class="skip-link screen-reader-text" href="#site-navigation"><?php esc_attr_e( 'Skip to navigation', 'ipress' ); ?></a>
+<a class="skip-link screen-reader-text" href="#content"><?php esc_attr_e( 'Skip to content', 'ipress' ); ?></a>